Add Grafana to Kubernetes cluster
Deploy Grafana to our Kubernetes cluster. We write a very simple set of
manifests from scratch, which will go on to form a Helm chart.

For now, we hardcode a plain text password, but we'll want a better
strategy in place before deploying this to the public internet.

This commit only contains a WIP set of dashboards. We'll determine the
rest once we've deployed Grafana to production and can start playing
with production data.