Add initial grafana graphs for SLO
We want to measure metrics relating to our SLO.

For now, I've included graphs for our service level indicators, which
also show when it is out of the SLO threshold in the short term.

I still want to add an error budget burn down chart, but I need to
define some Prometheus recording rules before I can do so.