Add POC Availability Error Budget burn rate chart
Proof of concept for what the availability error budget burn rate chart could look
like. Its not yet entirely clear to me how I'll replicate this chart for latency.
Perhaps we can do # of requests that take longer than the .99 threshold
over the last 30 days... (although we'll have to make sure those
actually use different time windows).

I also need to finalize the windows over which I plan to track the error
budget, and when it resets and what not (i.e. is it on a month to month
basis, does it reset every 7 days, etc...)