Add Prometheus Rules and Alertmanager
Follow the instructions in
https://github.com/coreos/prometheus-operator/blob/master/Documentation/user-guides/alerting.md
to add Alertmanager. Additionally, add PrometheusRules for monitoring
our SLO.

For now, we use fairly primitive SLO monitoring (basically alert if our
SLO is violated for more than X minutes). Long term, it would be great
to monitor on error rate burn down, but I'm not sure I'm comfortable
enough with Prometheus yet.

Deploying Alertmanager requires us to specify a k8s Secret. For now,
we've defined the Secret as a file which we keep out of source control
by using a `.sample` file and `envsubst`. We should reconsider this once
we have a better secrets solution.

Alertmanager is configured to raise alerts via Opsgenie, on which I've
created a free account.

Note: We had to upgrade the version of the prometheus operator we use in
production.