Add k8s resource usage instrumentation/monitoring
As we look to minimize the amount of money we spend on our Kubernetes
cluster, we want a good sense of what percentage of our k8s cluster
resources we're actually utilizing.

We obtain this through two mechanisms. First, we set up
[metrics-server](https://github.com/kubernetes-incubator/metrics-server),
so that `kubectl top` will work. metrics-server will also be useful for
when we add HPA.

Additionally, we configure Prometheus to scape kubelet/cadvisor metrics.
These metrics will give us a finer grained sense of resource
usage/allocation on our Kubernetes cluster.

The README.md files have descriptions for where we obtained source code.