Fix blog ServiceMonitor w/ new helm labelling
I specified that we should monitor the blog endpoint via labels.
However, I changed the labels we use for the blog when we switched to
the new helm based deploy in the previous commit. Update the
ServiceMonitor to use the correct labels. I verified the Prometheus is
successfully scraping metrics again.