Improve secret flow grafana
Improve how we manage secrets throughout the repo by using helm
templates for the base64 encoding.

Additionally, use this new secret pattern for grafana, so that we no longer store the
admin password in source control (not that it really mattered because
grafana wasn't publicly accessible).

I will switch to using base64 encoding for alertmanager in a separate
commit.