Improve documentation secret workflow alertmanager
Clarify that alertmanager only pre-encodes the secret as base64 before
placing in secret-values.yaml because the secret is an entire multiline
string, which is posing real difficult to embed otherwise.

Long term, something like
https://github.com/helm/helm/blob/master/docs/chart_template_guide/accessing_files.md
might be a better solution... but I"m worried about how that will
conform with the `secret-values.yaml` interface.