Update non-HA applications with HA tolerances
Because we've tainted our spot instance nodes, our non-HA applications
need to specify a toleration if they are ok running on spot instances.
Add the necessary taints to grafana, metrics-server, next-cloud, and
prometheus-operator.

Additionally, reduce the number of replicas we run for some of our
non-HA applications (alertmanager and prometheus), as we can't currently
make a good case for needing more than one replica.