Schedule updates for k8s cluster
Include a design doc on how I'll handle cluster upgrades and upgrades for
applications running on the cluster.

To support the latter, ensure we specify tag for all image versions,
so I'm not relying on any implicit versions. In other words, we control
all application versions running on our cluster.