Add LivenessProbe and ReadinessProbe for blog
Add a LivenessProbe and ReadinessProbe for my blog application. Based on
querying the `/metrics` endpoint, because as of now, we don't have a
true `status` or `health` endpoint.

For now, we only add these probes to the blog, as it is the application
we deploy the most and our most "critical" application. Its trivial to
add to our other pods should we want to.